+++ sdk/storage/azure-storage-files-shares/test/ut/share_file_client_test.cpp
@@ -370,7 +370,7 @@ namespace Azure { namespace Storage { namespace Test {
   {
     CHECK_SKIP_TEST();
 
-    const auto blobContent = RandomBuffer(8_MB);
+    const auto blobContent = RandomBuffer(static_cast<size_t>(8_MB));
 
     auto testUploadFromBuffer = [&](int concurrency,
                                     int64_t bufferSize,
@@ -388,10 +388,12 @@ namespace Azure { namespace Storage { namespace Test {
       }
 
       auto fileClient = m_shareClient->GetRootDirectoryClient().GetFileClient(RandomString());
-      EXPECT_NO_THROW(fileClient.UploadFrom(blobContent.data(), bufferSize, options));
-      std::vector<uint8_t> downloadBuffer(bufferSize, '\x00');
+      EXPECT_NO_THROW(
+          fileClient.UploadFrom(blobContent.data(), static_cast<size_t>(bufferSize), options));
+      std::vector<uint8_t> downloadBuffer(static_cast<size_t>(bufferSize), '\x00');
       fileClient.DownloadTo(downloadBuffer.data(), downloadBuffer.size());
-      std::vector<uint8_t> expectedData(blobContent.begin(), blobContent.begin() + bufferSize);
+      std::vector<uint8_t> expectedData(
+          blobContent.begin(), blobContent.begin() + static_cast<size_t>(bufferSize));
       EXPECT_EQ(downloadBuffer, expectedData);
     };
 
@@ -412,13 +414,16 @@ namespace Azure { namespace Storage { namespace Test {
 
       const std::string tempFileName = RandomString();
       WriteFile(
-          tempFileName, std::vector<uint8_t>(blobContent.begin(), blobContent.begin() + fileSize));
+          tempFileName,
+          std::vector<uint8_t>(
+              blobContent.begin(), blobContent.begin() + static_cast<size_t>(fileSize)));
       auto fileClient = m_shareClient->GetRootDirectoryClient().GetFileClient(RandomString());
       EXPECT_NO_THROW(fileClient.UploadFrom(tempFileName, options));
       DeleteFile(tempFileName);
-      std::vector<uint8_t> downloadBuffer(fileSize, '\x00');
+      std::vector<uint8_t> downloadBuffer(static_cast<size_t>(fileSize), '\x00');
       fileClient.DownloadTo(downloadBuffer.data(), downloadBuffer.size());
-      std::vector<uint8_t> expectedData(blobContent.begin(), blobContent.begin() + fileSize);
+      std::vector<uint8_t> expectedData(
+          blobContent.begin(), blobContent.begin() + static_cast<size_t>(fileSize));
       EXPECT_EQ(downloadBuffer, expectedData);
     };
